examples/tool-agent/python/requirements.txt
anthropic>=0.40
fastapi>=0.115
uvicorn[standard]>=0.32
asyncpg>=0.29
pydantic>=2.7
pyjwt>=2.9
langfuse>=2.50
opentelemetry-instrumentation-anthropic>=0.30
